requirements.txt
pycryptodome
memory-profiler